requirements.txt
pycryptodome
python-socketio
python-socketio[client]
eventlet
